Pipeline implementation
Add Pipeline and PipelineRun definition.
Since TaskRuns do not work because of https://github.com/knative/build-pipeline/issues/403,
I added a PVC yaml to preprovision the PVC. As long as it has the
right name, the TaskRun picks it up, which is a decent workaround.
When the TaskRun is deleted, the PVC stays there too.

NOTE: This does not work as it is! Main open issues:
- kaniko is missing the credentials to push images
- the pipeline run does nothing