Add readiness probe to the DB container
Move the readiness and schema upgrade from init containers in the
API pod to the database side so that the database service is self
contained and the DB pod readiness matches the actual DB service
readiness.

This is implemented using a feature of the postgres docker image
that permits running init scripts when the DB is started and
listening to socket but not yet via the network interface.
The docker image starts the DB, runs the migration scripts mounted
from the git repo and then starts listening on the network
interface. The readinessProbe and livenessProbe check the database
readiness via the cluster DB hostname and port.

The database docker image has been refactored to be as thin as
possible.