Optimize the API docker image
Use the strech-slim base package to reduce size. Do not use Alpine
because there are no wheels for pandas and it takes a long time to
build it.
Remove build time dependencies after install. Remove or clean-up
caches for apt-get and pip.
Sort things so that it is less likely to invalidate layers when
making code changes. Enabled building from a custom git URL/ref.