Support pre-provisioned DB
Support a preprovisioned DB in the helm chart. The DB can be
pg or mysql as well. Add helm templates to calculate the right
DB connection string parts, depending on whether the
preprovisioned DB is used or not.

Extend the helm deploy task to load values from a file, so that
values may be optional, since tekton does not support
conditions in tasks / pipelines yet.